feat: modified files to allow  eslint to not throw errors
Postcss.json and coverage were causing errors since they were not being found in tsconfig while . I modified tsconfig.json to include  postcss.json and created a .eslintignore in order to not go over coverage. This is because we should not be modifying coverage as it will be generated at runtime.